pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["fiftinmen <fiftin.men@gmail.com>"]
readme = "README.md"
packages = [
    { include = "task_manager" },
]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
dj-database-url = "^2.2.0"
python-dotenv = "^1.0.1"
django-bootstrap5 = "^24.2"
whitenoise = "^6.7.0"
psycopg2-binary = "^2.9.9"
gunicorn = "^22.0.0"
django-extensions = "^3.2.1"
django-admin = "^2.0.2"
django-filter = "^24.2"
rollbar = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
flake8 = "^7.1.0"
django-admin = "^2.0.2"
pytest-django = "^4.5.2"
ipython = "^8.5.0"
django-shell-plus = "^1.1.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
